+++ internal/cmd/sync/cmd.go
@@ -88,7 +88,7 @@ func Cmd() *cobra.Command {
 				DryRun:            dryRun,
 				IgnoreUncommitted: None[configdomain.IgnoreUncommitted](),
 				Order:             None[configdomain.Order](),
-				ProposeHeadless:   None[forgedomain.ProposeHeadless](),
+				ProposeHeadless:   None[configdomain.ProposeHeadless](),
 				PushBranches:      pushBranches,
 				Stash:             None[configdomain.Stash](),
 				Verbose:           verbose,
