+++ internal/config/configdomain/propose_headless.go
@@ -1,4 +1,4 @@
-package forgedomain
+package configdomain
 
 // ProposeHeadless indicates whether the propose command should skip opening a browser.
 type ProposeHeadless bool
