+++ internal/config/configfile/save_test.go
@@ -75,7 +75,7 @@ func TestSave(t *testing.T) {
 				PerennialRegex:              perennialRegex,
 				ProposalBreadcrumb:          Some(configdomain.ProposalBreadcrumbBranches),
 				ProposalBreadcrumbDirection: Some(configdomain.ProposalBreadcrumbDirectionUp),
-				ProposeHeadless:             Some(forgedomain.ProposeHeadless(true)),
+				ProposeHeadless:             Some(configdomain.ProposeHeadless(true)),
 				PushBranches:                Some(configdomain.PushBranches(true)),
 				PushHook:                    Some(configdomain.PushHook(true)),
 				ShareNewBranches:            Some(configdomain.ShareNewBranchesPropose),
