+++ internal/forge/forgedomain/connector.go
@@ -24,7 +24,6 @@ type Connector interface {
 type CreateProposalArgs struct {
 	Branch         gitdomain.LocalBranchName
 	FrontendRunner subshelldomain.Runner
-	Headless       ProposeHeadless
 	MainBranch     gitdomain.LocalBranchName
 	ParentBranch   gitdomain.LocalBranchName
 	ProposalBody   Option[gitdomain.ProposalBody]
