+++ internal/forge/gitlab/web_connector.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/git-town/git-town/v22/internal/browser"
 	"github.com/git-town/git-town/v22/internal/browser/browserdomain"
+	"github.com/git-town/git-town/v22/internal/config/configdomain"
 	"github.com/git-town/git-town/v22/internal/forge/forgedomain"
 	"github.com/git-town/git-town/v22/internal/messages"
 	"github.com/git-town/git-town/v22/internal/subshell/subshelldomain"
@@ -20,7 +21,8 @@ var (
 // WebConnector provides connectivity to GitLab through the browser.
 type WebConnector struct {
 	forgedomain.HostedRepoInfo
-	browser Option[browserdomain.Browser]
+	browser  Option[browserdomain.Browser]
+	headless configdomain.ProposeHeadless
 }
 
 func (self WebConnector) BrowseRepository(runner subshelldomain.Runner) error {
@@ -30,7 +32,7 @@ func (self WebConnector) BrowseRepository(runner subshelldomain.Runner) error {
 
 func (self WebConnector) CreateProposal(data forgedomain.CreateProposalArgs) error {
 	proposalURL := self.NewProposalURL(data)
-	if data.Headless {
+	if self.headless {
 		fmt.Printf(messages.BrowserOpen, proposalURL)
 	} else {
 		browser.Open(proposalURL, data.FrontendRunner, self.browser)
