+++ internal/forge/glab/connector.go
@@ -5,6 +5,7 @@ import (
 	"strings"
 
 	"github.com/git-town/git-town/v22/internal/cli/print"
+	"github.com/git-town/git-town/v22/internal/config/configdomain"
 	"github.com/git-town/git-town/v22/internal/forge/forgedomain"
 	"github.com/git-town/git-town/v22/internal/forge/gitlab"
 	"github.com/git-town/git-town/v22/internal/git/gitdomain"
@@ -26,6 +27,7 @@ type Connector struct {
 	Backend  subshelldomain.Querier
 	Frontend subshelldomain.Runner
 	Log      print.Logger
+	headless configdomain.ProposeHeadless
 }
 
 // ============================================================================
@@ -53,7 +55,7 @@ func (self Connector) CreateProposal(data forgedomain.CreateProposalArgs) error 
 	if !hasTitle || !hasBody {
 		args = append(args, "--fill")
 	}
-	if data.Headless {
+	if self.headless {
 		args = append(args, "--yes")
 	} else {
 		args = append(args, "--web")
